backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
httpx==0.27.0
pdfplumber==0.11.0
python-dotenv==1.0.1
